pyproject.toml
[project]
name = "learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "fastapi==0.104.1",
    "google-generativeai==0.3.2",
    "orjson==3.9.10",
    "passlib[bcrypt]==1.7.4",
    "pdfplumber==0.10.3",
    "pypdf2==3.0.1",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "qdrant-client==1.7.0",
    "rich==13.7.0",
    "together==1.2.11",
    "uvicorn[standard]==0.24.0",
]
